backend/requirements.txt
fastapi==0.115.11
openai==1.68.0
psycopg2==2.9.9
pydantic==2.10.6
python-dotenv==1.0.1
tabulate==0.9.0
uvicorn==0.34.0